Empezando grupos

+++ app/src/main/java/com/example/crycast/model/GroupsUsers.kt
@@ -4,6 +4,6 @@ import androidx.room.Entity
 
 @Entity(primaryKeys = ["groupId", "userId"])
 data class GroupsUsers(
-    var groupId: Int,
+    var groupId: Long,
     var userId: Int
 )
